Move tests about to make more sense
