+++ api/tests/test_teledeclaration.py
@@ -795,41 +795,3 @@ class TestTeledeclarationApi(APITestCase):
             Teledeclaration.objects.bulk_create([teledeclaration_1, teledeclaration_2])
         except IntegrityError:
             self.fail("Should be able to have several submitted TDs for deleted canteens")
-
-    @override_settings(HOSTNAME="ma-cantine.example.org")
-    @override_settings(ENABLE_TELEDECLARATION=True)
-    @freeze_time("2021-01-20")
-    @authenticate
-    def test_authentication_method_created_with_api(self):
-        """
-        Teledeclarations created via a third party API should get an authentication method of 'API'
-        """
-        diagnostic = DiagnosticFactory.create(year=2020)
-        diagnostic.canteen.managers.add(authenticate.user)
-        payload = {"diagnosticId": diagnostic.id}
-
-        self.client.credentials(HTTP_HOST="third-party")
-        self.client.post(reverse("teledeclaration_create"), payload)
-
-        diagnostic.refresh_from_db()
-        td = diagnostic.teledeclaration_set.first()
-        self.assertEqual(td.history.first().authentication_method, "API")
-
-    @override_settings(HOSTNAME="ma-cantine.example.org")
-    @override_settings(ENABLE_TELEDECLARATION=True)
-    @freeze_time("2021-01-20")
-    @authenticate
-    def test_authentication_method_created_with_website(self):
-        """
-        Teledeclarations created via the website should get an authentication method of 'WEBSITE'
-        """
-        diagnostic = DiagnosticFactory.create(year=2020)
-        diagnostic.canteen.managers.add(authenticate.user)
-        payload = {"diagnosticId": diagnostic.id}
-
-        self.client.credentials(HTTP_HOST="ma-cantine.example.org")
-        self.client.post(reverse("teledeclaration_create"), payload)
-
-        diagnostic.refresh_from_db()
-        td = diagnostic.teledeclaration_set.first()
-        self.assertEqual(td.history.first().authentication_method, "WEBSITE")
